backend/requirements.txt
Django
djangorestframework
django-cors-headers
djoser
djangorestframework-simplejwt
Pillow
gunicorn
psycopg2-binary
dj-database-url
whitenoise